General UI Stuff
Instead of having the UI elements scale to the size of the page, the page itself is now no longer resizeable. Added functionality to some home buttons and buttons on the home page. The inventory page causes an error when trying to return to the home page and i don't know how to fix it. Changed the appearance of the home page to be more in line with the rest of the pages